Moved SQL-related classes to a new folder.
