+++ src/java/me/rishshadra/nhstracker/GroupHourHandler.java
@@ -5,6 +5,7 @@
  */
 package me.rishshadra.nhstracker;
 
+import me.rishshadra.nhstracker.sql.Reader;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.SQLException;
