+++ src/java/me/rishshadra/nhstracker/Student.java
@@ -5,6 +5,7 @@
  */
 package me.rishshadra.nhstracker;
 
+import me.rishshadra.nhstracker.sql.Reader;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.logging.Level;
