+++ src/java/me/rishshadra/nhstracker/sql/Database.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package me.rishshadra.nhstracker;
+package me.rishshadra.nhstracker.sql;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
