+++ src/java/me/rishshadra/nhstracker/utils/DBCleanup.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import me.rishshadra.nhstracker.Reader;
+import me.rishshadra.nhstracker.sql.Reader;
 import me.rishshadra.nhstracker.Student;
 
 /**
