+++ src/java/me/rishshadra/nhstracker/utils/DatabaseStatus.java
@@ -17,7 +17,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import me.rishshadra.nhstracker.Database;
+import me.rishshadra.nhstracker.sql.Database;
 
 /**
  *
