+++ src/java/me/rishshadra/nhstracker/utils/IDMatcher.java
@@ -13,8 +13,8 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import me.rishshadra.nhstracker.Database;
-import me.rishshadra.nhstracker.Reader;
+import me.rishshadra.nhstracker.sql.Database;
+import me.rishshadra.nhstracker.sql.Reader;
 import me.rishshadra.nhstracker.Student;
 
 /**
