+++ test/me/rishshadra/nhstracker/DatabaseTest.java
@@ -5,6 +5,7 @@
  */
 package me.rishshadra.nhstracker;
 
+import me.rishshadra.nhstracker.sql.Database;
 import java.sql.Connection;
 import org.junit.After;
 import org.junit.AfterClass;
